Create a repository implementation on retrofit, and left local repository
