+++ app/src/main/java/by/yancheuski/githubusers/data/UsersRepoImpl.kt
@@ -1,0 +1,23 @@
+package by.yancheuski.githubusers.data
+
+import android.os.Handler
+import android.os.Looper
+import by.yancheuski.githubusers.domain.entities.UserEntity
+import by.yancheuski.githubusers.domain.repos.UsersRepo
+
+private const val DATA_LOADING_FAKE_DELAY = 3_000L
+
+class UsersRepoImpl : UsersRepo {
+
+    private val data: List<UserEntity> = listOf(
+        UserEntity("mojombo", 1, "https://avatars.githubusercontent.com/u/1?v=4"),
+        UserEntity("defunkt", 2, "https://avatars.githubusercontent.com/u/2?v=4"),
+        UserEntity("pjhyett", 3, "https://avatars.githubusercontent.com/u/3?v=4"),
+    )
+
+    override fun getUsers(onSuccess: (List<UserEntity>) -> Unit, onError: ((Throwable) -> Unit)?) {
+        Handler(Looper.getMainLooper()).postDelayed({
+            onSuccess(data)
+        }, DATA_LOADING_FAKE_DELAY)
+    }
+}
