+++ app/src/main/java/by/yancheuski/githubusers/data/UsersRetrofitImpl.kt
@@ -1,23 +1,42 @@
 package by.yancheuski.githubusers.data
 
-import android.os.Handler
-import android.os.Looper
 import by.yancheuski.githubusers.domain.entities.UserEntity
+import by.yancheuski.githubusers.domain.repos.GitHubUsersApi
 import by.yancheuski.githubusers.domain.repos.UsersRepo
-
-private const val DATA_LOADING_FAKE_DELAY = 3_000L
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
 
 class UsersRetrofitImpl : UsersRepo {
 
-    private val data: List<UserEntity> = listOf(
-        UserEntity("mojombo", 1, "https://avatars.githubusercontent.com/u/1?v=4"),
-        UserEntity("defunkt", 2, "https://avatars.githubusercontent.com/u/2?v=4"),
-        UserEntity("pjhyett", 3, "https://avatars.githubusercontent.com/u/3?v=4"),
-    )
+    private val baseUrl = "https://api.github.com/"
+
+    private val retrofit = Retrofit.Builder()
+        .baseUrl(baseUrl)
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+
+    private var usersApi: GitHubUsersApi = retrofit.create(GitHubUsersApi::class.java)
 
     override fun getUsers(onSuccess: (List<UserEntity>) -> Unit, onError: ((Throwable) -> Unit)?) {
-        Handler(Looper.getMainLooper()).postDelayed({
-            onSuccess(data)
-        }, DATA_LOADING_FAKE_DELAY)
+        usersApi.getListOfUsers().enqueue(object : Callback<List<UserEntity>> {
+            override fun onResponse(
+                call: Call<List<UserEntity>>,
+                response: Response<List<UserEntity>>,
+            ) {
+                val body = response.body()
+                if (response.isSuccessful && body != null) {
+                    onSuccess.invoke(body)
+                } else {
+                    onError?.invoke(Throwable())
+                }
+            }
+
+            override fun onFailure(call: Call<List<UserEntity>>, t: Throwable) {
+                onError?.let { it(t) }
+            }
+        })
     }
 }
