+++ app/src/main/java/by/yancheuski/githubusers/domain/repos/GitHubUsersApi.kt
@@ -3,10 +3,9 @@ package by.yancheuski.githubusers.domain.repos
 import by.yancheuski.githubusers.domain.entities.UserEntity
 import retrofit2.Call
 import retrofit2.http.GET
-import retrofit2.http.Query
 
 interface GitHubUsersApi {
 
-    @GET("users/")
-    fun getListOfUsers(@Query("login") login: String): Call<UserEntity>
+    @GET("users")
+    fun getListOfUsers(): Call<List<UserEntity>>
 }
